+++ src/index.js
@@ -1,17 +1,4 @@
 import './style.css'
 import { app } from './layouts/app'
 
-
-function moduleRender() {
-	const jsTarget = document.createElement('div')
-	jsTarget.classList.add('js-target')
-
-	// modules to load below
-	jsTarget.appendChild(app.layoutElement)
-	
-
-	return jsTarget
-}
-
-document.body.appendChild(moduleRender())
-app.addProjectTabListener()
+document.body.appendChild(app.layoutElement)
