+++ src/style.css
@@ -11,7 +11,7 @@ p {
   margin: 0;
 }
 
-.layout {
+.app {
   margin: 0 auto;
   width: min-content;
   display: grid;
